Remove autoconf support
Summary:
Currently, we have a mixed build system with some autoconf support, and some cmake support. Maintaining
this duplication is a lot of overhead.  Meanwhile, autoconf as a whole is archaic and hard to maintain.
There is little to no value with maintaining autoconf support -- it primarily exists to continue to
support qt builds, which also should be removed.

Test Plan:
Build and run bitcoind

Tags: #build