Fix bench-bitcoin bug in BlockToJsonVerbose benchmark. (#102)
62c2bc5 added the requirement to check LowS for verbose serialization of sigs, but we didn't fix the benchmarks.

Since we always enforce Low S, the signature printing (for the human readable script asm) needs to check if the signature has low S. For that, we need access to a secp256k1 verify context. ECCVerifyHandle sets this up for us in the constructor and tears it down again in the destructor.