Make non-pushonly scriptsig in coinbase inputs standard
Currently, non-pushonly scriptSigs are non-standard; but mining pool software relies on the scriptSig of the coinbase being arbitray.
We enforce standardness on all txs, so blocks having non-pushonly coinbases scriptSigs will be rejected, which is not what we want.

This will remove the "scriptsig-not-pushonly" restriction for standardness on coinbase inputs.