Update default ports to be based off p2p port 10605
Since Lotus is not a ledger fork, there is no reason to maintain the
same default ports. In fact, this makes it harder for individuals to run
nodes on the same server as whatever of flavor of bitcoind they prefer.
This we're going to move to a different port to support ease of use for
node operators.

New ports:
* mainnet
  * p2p - 10605
  * rpc - 10604
* testnet
  * p2p - 11605
  * rpc - 11604
* regtest
  * p2p - 12605
  * rpc - 12604

Closes #153