Fix several mistakes in gitian/artifact bundling due to binary rename (#177)
In the zeal to rename everything to lotus, some changes to the gitian
linux and windows scripts were erroneously included. This commit reverts
those changes. Additionally, a few icon names were not correctly renamed
in the scripts despite the files being renamed. Longer term, more
references to bitcoin should be swapped to lotus to avoid confusion, but
within these scripts the references are immaterial.