Add `enableMinerFund` to config to enable testing
This commit adds `enableMinerFund` to the config so that it can be
overriden in w/o setting gArgs. It also adds flags to the
`BlockAssembler` options as well as the validation options in order to
support passing these configs down to the appropriate code. However,
current testing setup relies on the global `GetConfig()` singleton. This
singleton always returns the chainparams from the global `Params()`
singleton function. The testing setup changes the Params after
initialization, thus we need to ensure that `EnableMinerFund()` returns
the appropriate default consensus parameter if it has not been formally
set by the command line or the testing setup. In the future, global
singletons should be removed so that tests can be ran in parallel.
However, this requires substantial refactors to the entire codebase to
ensure proper initialization ordering and encapsulation of
functionality.