Implement a new simplified mining API `getunsolvedblock`
This API takes an address and returns a target, and the serialized block
as hex. Because the nonce was expanded, there is no need to grind on the
coinbase in order to produce a valid block. Thus, we can implement a
miner which takes only the block header. However, currently we also
include the transactions as there is no internal caching for the
transaction set. These are needed in submit block for validating the
merkle root and broadcasting the transactions later. Mining software
will need to hold on to them, but does not need to do anything with
them.