Ensure rpc_generateblock always passes
While redenominating the code to 10^6 from 10^8, this particular
spot was missed. The test continued to pass because `listunspent` is
nondeterministic. Sometimes it was pulling a tx with sufficient funds,
and sometimes it was creating transactions that overspent. This change
ensures that both transactions that can be used in subsequent lines
have enough funds to pass the test.