Implement random selection of output addresses based on epoch block (#181)
We've decided to implement a lottocracy for the system of governance on
Lotus. As such, each weak (epoch) a set of random citizens will be
selected to receive a portion of the block reward which they may
allocate however they want. This code implements the random selection
from a list of addresses of verified citizens. This is subject to modulo
bias, however the effect is miner given the use case. In the future,
better algorithms may be implemented.