Fix commitment in XAddress check byte hash calculation (#212)
Due to using CHashWriter, we were including VarInt sizes for the
two vectors which were included. This was a mistake and not within
the specification. This commit adds some extra checks, changes the
hash generation function, and introduces a second check to enable
legacy addresses to pass decoding for the time being.