Update seeds lists (#295)
This was manually updated in several places. The seeders are strangly
not seeding a variety of nodes. There are only a few available. In
addition to the nodes which the seeder is dumping, I have manually
updated the list with peers from my own node, and the node which has
been running the explorer.